Extract upgrade/details to pkg/upgrade/details for cross-repo sharing
Moves the UpgradeDetails type and its supporting machinery (State, Metadata,
observer pattern, mutation methods) from internal/ to pkg/ so external consumers
like Horde can import it without depending on Agent internals.

The only change to the package content is replacing the elastic-agent-libs
DownloadRate type with plain float64 — Agent already normalises +Inf/NaN in
SetDownloadProgress before storing, so the custom MarshalJSON is redundant.

Part of elastic/ingest-dev#7601 (item 9).
