+++ internal/pkg/agent/application/upgrade/marker_watcher_test.go
@@ -18,8 +18,8 @@ import (
 
 	"gopkg.in/yaml.v2"
 
-	"github.com/elastic/elastic-agent/internal/pkg/agent/application/upgrade/details"
 	"github.com/elastic/elastic-agent/pkg/core/logger/loggertest"
+	"github.com/elastic/elastic-agent/pkg/upgrade/details"
 )
 
 func TestMarkerWatcher(t *testing.T) {
