+++ testing/upgradetest/upgrader.go
@@ -20,10 +20,10 @@ import (
 	"github.com/otiai10/copy"
 
 	"github.com/elastic/elastic-agent/internal/pkg/acl"
-	"github.com/elastic/elastic-agent/internal/pkg/agent/application/upgrade/details"
 	v1client "github.com/elastic/elastic-agent/pkg/control/v1/client"
 	v2proto "github.com/elastic/elastic-agent/pkg/control/v2/cproto"
 	atesting "github.com/elastic/elastic-agent/pkg/testing"
+	"github.com/elastic/elastic-agent/pkg/upgrade/details"
 	"github.com/elastic/elastic-agent/pkg/version"
 )
